requirements_question1.txt
pyserial
numpy
Pillow
matplotlib


